+++ src/Portenta_H7_AsyncHTTPRequest.h
@@ -4,9 +4,7 @@
   For Portenta_H7 (STM32H7) with Vision-Shield Ethernet or Murata WiFi
   
   Portenta_H7_AsyncHTTPRequest is a library for the Portenta_H7 with with Vision-Shield Ethernet Ethernet or Murata WiFi
-  
-  Portenta_H7_AsyncHTTPRequest is a library for Portenta_H7
-  
+   
   Based on and modified from asyncHTTPrequest Library (https://github.com/boblemaire/asyncHTTPrequest)
   
   Built by Khoi Hoang https://github.com/khoih-prog/Portenta_H7_AsyncHTTPRequest
@@ -19,13 +17,14 @@
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
   You should have received a copy of the GNU General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>.  
  
-  Version: 1.2.0
+  Version: 1.3.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0    K Hoang     13/10/2020 Initial coding for Portenta_H7
   1.1.0    K Hoang     30/12/2021 Fix `multiple-definitions` linker error
   1.2.0    K Hoang     24/01/2022 Enable compatibility with old code to include only AsyncHTTPRequest_Generic.h
+  1.3.0    K Hoang     02/09/2022 Fix bug. Improve debug messages. Optimize code
  *****************************************************************************************************************************/
 
 #pragma once