+++ src/Portenta_H7_AsyncHTTPRequest.hpp
@@ -4,9 +4,7 @@
   For Portenta_H7 (STM32H7) with Vision-Shield Ethernet or Murata WiFi
   
   Portenta_H7_AsyncHTTPRequest is a library for the Portenta_H7 with with Vision-Shield Ethernet Ethernet or Murata WiFi
-  
-  Portenta_H7_AsyncHTTPRequest is a library for Portenta_H7
-  
+   
   Based on and modified from asyncHTTPrequest Library (https://github.com/boblemaire/asyncHTTPrequest)
   
   Built by Khoi Hoang https://github.com/khoih-prog/Portenta_H7_AsyncHTTPRequest
@@ -19,13 +17,14 @@
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
   You should have received a copy of the GNU General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>.  
  
-  Version: 1.2.0
+  Version: 1.3.0
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
   1.0.0    K Hoang     13/10/2020 Initial coding for Portenta_H7
   1.1.0    K Hoang     30/12/2021 Fix `multiple-definitions` linker error
   1.2.0    K Hoang     24/01/2022 Enable compatibility with old code to include only AsyncHTTPRequest_Generic.h
+  1.3.0    K Hoang     02/09/2022 Fix bug. Improve debug messages. Optimize code
  *****************************************************************************************************************************/
 
 #pragma once
@@ -60,13 +59,13 @@
   #define SHIELD_TYPE           "Portenta_H7 WiFi"
 #endif
 
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION            "Portenta_H7_AsyncHTTPRequest v1.2.0"
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION            "Portenta_H7_AsyncHTTPRequest v1.3.0"
 
 #define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_MAJOR      1
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_MINOR      2
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_MINOR      3
 #define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_PATCH      0
 
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_INT        1002000
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_INT        1003000
 
 #include "Portenta_H7_AsyncTCP.h"
 
@@ -248,28 +247,24 @@ class AsyncHTTPRequest
 
     struct  URL 
     {
-      char*   scheme;
-      char*   user;
-      char*   pwd;
-      char*   host;
-      int     port;
-      char*   path;
-      char*   query;
-      char*   fragment;
+      char 		*buffer;
+      char 		*scheme;
+      char 		*host;
+      int 		port;
+      char 		*path;
+      char 		*query;
       
-      URL():  scheme(nullptr), user(nullptr), pwd(nullptr), host(nullptr),
-              port(80), path(nullptr), query(nullptr), fragment(nullptr)
+      URL():	buffer(nullptr), scheme(nullptr), host(nullptr),
+        			port(80), path(nullptr), query(nullptr)
       {};
-      
-      ~URL() 
+        
+      ~URL()
       {
+        SAFE_DELETE_ARRAY(buffer)
         SAFE_DELETE_ARRAY(scheme)
-        SAFE_DELETE_ARRAY(user)
-        SAFE_DELETE_ARRAY(pwd)
         SAFE_DELETE_ARRAY(host)
         SAFE_DELETE_ARRAY(path)
         SAFE_DELETE_ARRAY(query)
-        SAFE_DELETE_ARRAY(fragment)
       }
     };
 
@@ -316,6 +311,7 @@ class AsyncHTTPRequest
     size_t      available();                                            // response available
     size_t      responseLength();                                       // indicated response length or sum of chunks to date
     int         responseHTTPcode();                                     // HTTP response code or (negative) error code
+    String      responseHTTPString();
     String      responseText();                                         // response (whole* or partial* as string)
     
     char*       responseLongText();                                     // response long (whole* or partial* as string)